requirements.txt
# Core API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# LLM Providers
openai>=1.3.0
groq>=0.4.1

# Configuration and Environment
python-dotenv>=1.0.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data Processing
typing-extensions>=4.8.0
requests==2.31.0
